refactor: QueryRequest 및 QueryResponse를 QueryCommand 및 QueryResult로 대체
1. QueryRequest와 QueryResponse를 QueryCommand와 QueryResult로 변경한다.
2. Service 및 Facade 계층의 메서드 시그니처와 내부 로직을 수정한다.
3. Token 계산 및 QueryLog 생성 과정에서 새로운 객체 사용하도록 수정한다.

Resolves #47
